+++ cogless.py
@@ -1,0 +1,22 @@
+import discord
+import os
+
+from dotenv import load_dotenv
+load_dotenv()
+TOKEN = os.getenv('TOKEN')
+
+client = discord.Client()
+
+@client.event
+async def on_ready():
+    print(f'We have logged in as {client.user}')
+
+@client.event
+async def on_message(msg):
+    # print(msg.content)
+    if msg.author == client.user:
+        return
+    if msg.content.startswith('$hello'):
+        await msg.channel.send(f'Hello, {msg.author.name}!')
+
+client.run(TOKEN)
