+++ images/vscode_create_main_and_env_files.gif
