+++ main.py
@@ -1,22 +1,16 @@
-import discord
+# main.py
+from discord.ext import commands
 import os
-
 from dotenv import load_dotenv
+
 load_dotenv()
 TOKEN = os.getenv('TOKEN')
 
-client = discord.Client()
+client = commands.Bot(command_prefix = "!")
 
-@client.event
-async def on_ready():
-    print(f'We have logged in as {client.user}')
-
-@client.event
-async def on_message(msg):
-    # print(msg.content)
-    if msg.author == client.user:
-        return
-    if msg.content.startswith('$hello'):
-        await msg.channel.send(f'Hello, {msg.author.name}!')
+# Looks inside the /cogs/ folder and loads up all of our cogs
+for filename in os.listdir("./cogs"):
+	if filename.endswith(".py"):
+		client.load_extension("cogs." + filename[:-3])
 
 client.run(TOKEN)
