--- w1/discord_py_cheatsheet.md
@@ -1,6 +1,0 @@
-<h1 align="center">Discord.py CheatSheet</h1>
-
-## Events
-* Basic Structure of Event Callbacks
-```
-```
