--- workbook-template.md
@@ -1,28 +1,0 @@
-<h1 align="center">Participant's Workbook Template</h1>
-
-> Related Pages: [DecodED 3](./README.md) | [Participant's Handbook for Workshop 1: Intro to Discord.py](./workshop1/README.md)
-
-<TODO 0. add internal links to related pages, namely: the main README (DecodED3), the previous workbook and the next workshop>
-
----
-
-<h2>Table of Contents</h2>
-<details>
-<summary>Table of Contents</summary>
-
-
-</details>
-
-## Links
-
-
-## Part 1: 
-
-
-
-## Part 2: 
-
-
-
-## Part 3: 
-
